k9server/requirements/dev-requirements.txt
Django==1.8.2
decorator==3.4.2
django-ajaxutils==0.2
gnureadline==6.3.3
ipython==3.1.0
nltk==3.0.2
python-dateutil==2.4.2
simplejson==3.7.3
six==1.9.0
textblob==0.9.0
wsgiref==0.1.2
